pyproject.toml
[project]
name = "connection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-tasks>=0.9.0",
    "django-vite>=3.1.0",
    "environs[django]>=14.5.0",
    "heroicons[django]>=2.13.0",
    "mrml>=0.2.0",
    "mysqlclient>=2.2.7",
    "pillow>=12.0.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "coverage>=7.11.3",
    "django-browser-reload>=1.21.0",
    "django-debug-toolbar>=6.1.0",
    "django-watchfiles>=1.4.0",
    "factory-boy>=3.3.3",
    "playwright>=1.56.0",
]
